refactor: LottoFactory 디렉토리 생성 및 클래스 정리
