+++ src/main/java/lotto/domain/lottofactory/LottoFactory.java
@@ -1,4 +1,6 @@
-package lotto.domain.lotto;
+package lotto.domain.lottofactory;
+
+import lotto.domain.lotto.Lotto;
 
 public interface LottoFactory {
